move linebreak tests
